third_party/blink/web_tests/platform/win11-arm64/external/wpt/webtransport/stats.https.any.sharedworker-expected.txt

This is a testharness.js-based test.
[FAIL] WebTransport client should be able to provide droppedIncoming values for datagrams
  assert_greater_than: expected a number greater than 0 but got 0
Harness: the test ran to completion.
